voice-agent/pyproject.toml
[project]
name = "voice-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "groq>=1.0.0",
    "openai[voice-helpers]>=2.14.0",
    "python-dotenv>=1.2.1",
    "speechrecognition[audio]>=3.14.5",
]
